5.1 Routes
I rerouted my root page to the index page, although I'll probably end
up placing the carousel at the top of that page further down the road.

For "redirect", I made it so that a successful item submission goes to
the carousel 'landing_page', so they can see the item in the carousel.

I created links for the nav menu, using '_path' in each case since I
might change the URLs of the pages in the future.

I also used a dynamic link in the carousel using '<%= link_to
product.name, product %>' so that a user can visit the product's page
by clicking on the name when they see it in the carousel.