5.7 The Comment Model
Created the Comment model and declared the relationships to the User
and Product models by using the "relations" migration shortcut.
Set up two scopes for the Comment model to return a list of comments
ordered by rating from highest to lowest and vice versa.
Set up two custom methods for the Product model to return the comment
with the highest rating and the comment with the lowest rating.
Basic styling